Add support for parent/child documents
Previously, parent/child structures were not possible in Tire due to
needing additional {"_parent":"<id>"} metadata. This adds a check for
'_parent' in the mapping during bulk updating, and relies on foreign key
conventions to add the correct foreign key to use as the parent.

Querying parent/child could use better support, but this is the bare
minimum for using parent/child at all.